refactor(PieceDtoMapper): Domain <-> DTO 변환 로직 수정
